+++ tiger-common/src/main/java/de/gematik/rbellogger/util/GlobalServerMap.java
@@ -6,8 +6,10 @@ package de.gematik.rbellogger.util;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import org.assertj.core.util.VisibleForTesting;
 
 public class GlobalServerMap {
 
@@ -17,6 +19,9 @@ public class GlobalServerMap {
 
   private static final Map<Long, String> PROCESS_ID_TO_BUNDLED_SERVER_NAME = new HashMap<>();
   private static final ConcurrentMap<Integer, Long> PORT_TO_PROCESS_ID = new ConcurrentHashMap<>();
+
+  private static final ConcurrentMap<Integer, String> PORT_TO_LOCAL_SERVER_TYPES_NAMES =
+      new ConcurrentHashMap<>();
 
   public static void mapPortToProcessIds(int key, long value) {
     PORT_TO_PROCESS_ID.putIfAbsent(key, value);
@@ -33,5 +38,20 @@ public class GlobalServerMap {
 
   public static ConcurrentMap<Integer, Long> getPortToProcessId() {
     return PORT_TO_PROCESS_ID;
+  }
+
+  public static Optional<String> getServerNameForPort(int port) {
+    return Optional.ofNullable(PORT_TO_LOCAL_SERVER_TYPES_NAMES.get(port));
+  }
+
+  public static void addServerNameForPort(int port, String serverName) {
+    PORT_TO_LOCAL_SERVER_TYPES_NAMES.put(port, serverName);
+  }
+
+  @VisibleForTesting
+  public static void clear() {
+    PROCESS_ID_TO_BUNDLED_SERVER_NAME.clear();
+    PORT_TO_PROCESS_ID.clear();
+    PORT_TO_LOCAL_SERVER_TYPES_NAMES.clear();
   }
 }